Revert "Prefetch queries"
This reverts commit d11fe9bcdc650df3999ee4b33ec7f26ec6a0746e.
