+++ app/dashboard/total-dev-views/route.ts
@@ -1,8 +1,21 @@
+import type { DEVArticle } from "@/lib/types";
 import { NextResponse } from "next/server";
-import { getDevTotalViews } from "@/lib/dev";
 
 export const GET = async () => {
-  const totalViews = await getDevTotalViews();
+  const response = await fetch("https://dev.to/api/articles/me/published", {
+    headers: {
+      "api-key": process.env.DEV_API_KEY as string,
+    },
+    next: {
+      revalidate: 86400,
+    },
+  });
+  const data = (await response.json()) as DEVArticle[];
+
+  const totalViews = data.reduce(
+    (accumulator, currentValue) => accumulator + currentValue.page_views_count,
+    0
+  );
 
   return NextResponse.json({ count: totalViews });
 };
