+++ db/schema.ts
@@ -6,7 +6,6 @@ const conn = connect({
   host: process.env.DATABASE_HOST,
   username: process.env.DATABASE_USERNAME,
   password: process.env.DATABASE_PASSWORD,
-  fetch: (input, init) => fetch(input, { ...init, cache: "no-store" }),
 });
 
 export const db = drizzle(conn);
